# Files/swagger-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ShortStay API",
    "version": "1.0.0",
    "description": "\n## ShortStay Backend API\n\nA full-featured short-term property rental platform API.\n\n### Authentication\nMost endpoints require a JWT Bearer token. Include it in the `Authorization` header:\n```\nAuthorization: Bearer <your_token>\n```\n\n### Roles\n- **guest** — Can search properties, make bookings, write reviews, raise complaints\n- **host** — Can list/manage properties, manage bookings, receive payouts\n- **admin** — Full system access: user management, property approval, reports\n- **accountant** — Manages payments, refunds, disputes, payouts\n- **verifier** — Conducts property inspections and submits reports\n\n### Google OAuth Flow\n1. Frontend redirects browser to `GET /api/auth/google`\n2. After Google login, backend redirects to `{FRONTEND_URL}/auth/callback?token=...&role=...&user_id=...`\n3. Frontend stores the token from the URL query parameter\n      ",
    "contact": {
      "name": "ShortStay Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter the JWT token obtained from /api/auth/login"
      }
    },
...
  },
...
  "paths": {
...
    "/api/payments/process": {
      "post": {
        "summary": "Process a payment for a booking (guest only — non-Stripe flow)",
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "booking_id",
                  "payment_method"
                ],
                "properties": {
                  "booking_id": {
                    "type": "integer",
                    "example": 5
                  },
                  "payment_method": {
                    "type": "string",
                    "example": "card"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment processed"
          },
          "403": {
            "description": "Forbidden — guests only"
          }
        }
      }
    },
    "/api/payments/stripe/intent": {
      "post": {
        "summary": "Create a Stripe PaymentIntent (guest only)",
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Returns a `clientSecret` that the frontend uses with Stripe Elements to confirm payment.\n```js\nconst { clientSecret } = await api.post('/payments/stripe/intent', { booking_id });\n// Then use stripe.confirmCardPayment(clientSecret, { payment_method: ... })\n```\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "booking_id"
                ],
                "properties": {
                  "booking_id": {
                    "type": "integer",
                    "example": 5
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "PaymentIntent created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "clientSecret": {
                      "type": "string",
                      "example": "pi_3xxx_secret_yyy"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — guests only"
          }
        }
      }
    },
    "/api/payments/stripe/confirm": {
      "post": {
        "summary": "Confirm a Stripe payment after frontend completes card entry",
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "payment_intent_id"
                ],
                "properties": {
                  "payment_intent_id": {
                    "type": "string",
                    "example": "pi_3xxx"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment confirmed and booking updated"
          }
        }
      }
    },
    "/api/payments/retry/{payment_id}": {
      "post": {
        "summary": "Retry a failed payment (guest only)",
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "payment_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment retried"
          }
        }
      }
    },
    "/api/payments/my-payments": {
      "get": {
        "summary": "Get payment history for the logged-in user",
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of payments made by the user"
          }
        }
      }
    },
    "/api/payments/receipt/{booking_id}": {
      "get": {
        "summary": "Get payment receipt for a specific booking",
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Receipt data"
          },
          "404": {
            "description": "Receipt not found"
          }
        }
      }
    },
    "/api/payments/disputes": {
      "post": {
        "summary": "Raise a payment dispute (guest only)",
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "booking_id",
                  "reason"
                ],
                "properties": {
                  "booking_id": {
                    "type": "integer",
                    "example": 5
                  },
                  "reason": {
                    "type": "string",
                    "example": "Property did not match the description"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Dispute raised"
          },
          "403": {
            "description": "Forbidden — guests only"
          }
        }
      },
      "get": {
        "summary": "Get all disputes (accountant only)",
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of all disputes"
          },
          "403": {
            "description": "Forbidden — accountant only"
          }
        }
      }
    },
    "/api/payments/disputes/{id}/resolve": {
      "put": {
        "summary": "Resolve a dispute (accountant only)",
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Dispute ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "resolution"
                ],
                "properties": {
                  "resolution": {
                    "type": "string",
                    "example": "Refund approved after review"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Dispute resolved"
          }
        }
      }
    },
    "/api/payments/failed": {
      "get": {
        "summary": "Get all failed payments (accountant only)",
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of failed payments"
          }
        }
      }
    },
    "/api/payments/pending": {
      "get": {
        "summary": "Get all pending payments (accountant only)",
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of pending payments"
          }
        }
      }
    },
    "/api/payments": {
      "get": {
        "summary": "Get all payments (accountant only)",
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of all payments"
          }
        }
      }
    },
    "/api/payments/report": {
      "get": {
        "summary": "Generate payment report (accountant only)",
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Payment report data"
          }
        }
      }
    },
    "/api/payments/refund/{booking_id}": {
      "post": {
        "summary": "Issue a full refund for a booking (accountant only)",
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string",
                    "example": "Host cancelled last minute"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Refund issued"
          }
        }
      }
    },
    "/api/payments/refund/{booking_id}/partial": {
      "post": {
        "summary": "Issue a partial refund (accountant only)",
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "amount"
                ],
                "properties": {
                  "amount": {
                    "type": "number",
                    "example": 50
                  },
                  "reason": {
                    "type": "string",
                    "example": "Partial service failure"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Partial refund issued"
          }
        }
      }
    },
    "/api/payments/status/{payment_id}": {
      "put": {
        "summary": "Manually update payment status (accountant only)",
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "payment_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "pending",
                      "completed",
                      "failed",
                      "refunded"
                    ],
                    "example": "completed"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status updated"
          }
        }
      }
    },
    "/api/payments/reports/monthly": {
      "get": {
        "summary": "Monthly revenue report (accountant only)",
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Monthly revenue data"
          }
        }
      }
    },
    "/api/payments/reports/by-property": {
      "get": {
        "summary": "Revenue breakdown by property (accountant only)",
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Revenue per property"
          }
        }
      }
    },
    "/api/payments/reports/by-date": {
      "get": {
        "summary": "Revenue breakdown by date range (accountant only)",
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "start_date",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "end_date",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Revenue by date"
          }
        }
      }
    },
    "/api/payments/reports/refunds": {
      "get": {
        "summary": "Refund activity report (accountant only)",
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Refund data"
          }
        }
      }
    },
    "/api/payments/reports/host-payouts": {
      "get": {
        "summary": "Host payout report (accountant only)",
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Host payout history"
          }
        }
      }
    },
...
  }
}